Definition-typer til types-katalog
